Post.objects(user_id__in=user_ids).delete()

See merge request vodaplayvietnam/2025_ai_toktak_be!1146

+++ app/makers/docx.py
@@ -106,7 +106,17 @@ class DocxMaker:
         with open(txt_path, "w", encoding="utf-8") as f:
             f.writelines(txt_lines)
 
-        image_downloads = ImageMaker.save_normal_images(images, batch_id=batch_id)
+        image_downloads = []
+        check_image = {}
+        for image_url in images:
+            image_path = ImageMaker.save_image_url_get_path(image_url, batch_id)
+            if not image_path:
+                continue
+            if "toktak.ai" in image_url:
+                check_image[image_path] = True
+            else:
+                check_image[image_path] = False
+            image_downloads.append(image_path)
 
         with zipfile.ZipFile(zip_path, "w", zipfile.ZIP_DEFLATED) as zipf:
             # 🔍 Thêm file TXT nếu tồn tại
@@ -120,7 +130,10 @@ class DocxMaker:
         try:
             if os.path.isfile(txt_path):
                 os.remove(txt_path)
+
             for img_path in image_downloads:
+                if check_image.get(img_path):
+                    continue
                 if os.path.isfile(img_path):
                     os.remove(img_path)
